refactor(repl): extract REPL front-ends into @repo/repl package
The interpreter package carried modules that merely *consume* the
interpreter rather than being the language: the embeddable REPLs
(MemoryRepl/AgentRepl), the interactive CLI, and the shared-session
server. Move them into a new @repo/repl package so packages/interpreter
holds only the language + MCP integration.

- New packages/repl (@repo/repl): repl.ts, cli.ts, session-server.ts,
  with the repl / repl:attach scripts. Depends on @repo/interpreter.
- Repoint consumers: apps/mcp (MemoryRepl), apps/pi (AgentRepl),
  apps/lsp (session-server) now import from @repo/repl and declare the
  workspace dep. Root repl / repl:attach scripts target @repo/repl.
- Drop the moved files from the interpreter exports map, scripts, and
  the source.ts prompt-embed list.
- Move the REPL tests (repl-session, conversation-vars) into
  packages/repl/test. The three AgentRepl secret cases move out of
  secrets.test.ts into packages/repl/test/agent-repl-reset.test.ts:
  keeping them in the interpreter package would make it depend on
  @repo/repl, and turbo's `^typecheck` graph forbids that cycle.
- Register packages/repl in knip.json.
