+++ apps/lsp/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@repo/interpreter": "workspace:*",
+    "@repo/repl": "workspace:*",
     "vscode-languageserver": "^9.0.1",
     "vscode-languageserver-textdocument": "^1.0.12"
   },
