+++ package.json
@@ -10,8 +10,8 @@
     "typecheck": "turbo run typecheck",
     "test": "turbo run test",
     "test:watch": "turbo run test:watch",
-    "repl": "pnpm --filter @repo/interpreter repl",
-    "repl:attach": "pnpm --silent --filter @repo/interpreter repl:attach",
+    "repl": "pnpm --filter @repo/repl repl",
+    "repl:attach": "pnpm --silent --filter @repo/repl repl:attach",
     "lint": "biome ci",
     "format": "biome check --write",
     "knip": "knip",
