+++ packages/interpreter/src/source.ts
@@ -6,9 +6,6 @@ import { readFileSync } from "node:fs";
 const FILES = [
 	"arith.ts",
 	"lisp.ts",
-	"repl.ts",
-	"session-server.ts",
-	"cli.ts",
 	"grammar.ts",
 	"mcp.ts",
 	"mcp-broker.ts",
