+++ packages/interpreter/test/secrets.test.ts
@@ -4,7 +4,6 @@ import { join } from "node:path";
 import { fileURLToPath } from "node:url";
 import { afterAll, describe, expect, it } from "vitest";
 import { Interp, prelude, run, str } from "../src/lisp.ts";
-import { AgentRepl } from "../src/repl.ts";
 import { ev, freshInterp } from "./helpers.ts";
 
 const FIXTURE = fileURLToPath(
@@ -154,38 +153,6 @@ describe("secret registry (.env file loading)", () => {
 		const keys = ev("(secrets)", interp);
 		expect(keys).toContain("REPL_LINEAR_API_KEY");
 		expect(keys).not.toContain("NOT_A_SECRET");
-	});
-
-	it("re-injects explicitly-loaded secrets on AgentRepl reset()", () => {
-		const path = writeEnvFile("REPL_FOO=bar\n");
-		const repl = new AgentRepl();
-		repl.loadSecretsFromFile(path);
-		expect(repl.eval("(secrets)")).toContain("REPL_FOO");
-		repl.reset();
-		expect(repl.eval("(secrets)")).toContain("REPL_FOO");
-	});
-
-	it("keeps descriptions across an AgentRepl reset()", () => {
-		const repl = new AgentRepl();
-		repl.setSecrets({
-			REPL_FOO: { value: "bar", description: "the foo token" },
-		});
-		expect(repl.eval("(secrets)")).toContain("the foo token");
-		repl.reset();
-		expect(repl.eval("(secrets)")).toContain("the foo token");
-	});
-
-	it("does not auto-load $LISPTC_SECRETS_FILE for an embedded AgentRepl", () => {
-		const path = writeEnvFile("REPL_PI_TOKEN=t0ken\n");
-		const prev = process.env.LISPTC_SECRETS_FILE;
-		process.env.LISPTC_SECRETS_FILE = path;
-		try {
-			const repl = new AgentRepl();
-			expect(repl.eval("(secrets)")).not.toContain("REPL_PI_TOKEN");
-		} finally {
-			if (prev === undefined) delete process.env.LISPTC_SECRETS_FILE;
-			else process.env.LISPTC_SECRETS_FILE = prev;
-		}
 	});
 });
