+++ packages/repl/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@repo/repl",
+  "version": "0.0.0",
+  "type": "module",
+  "private": true,
+  "exports": {
+    "./repl.ts": "./src/repl.ts",
+    "./session-server.ts": "./src/session-server.ts",
+    "./cli.ts": "./src/cli.ts",
+    "./package.json": "./package.json"
+  },
+  "engines": {
+    "node": ">=22.6.0"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "repl": "node --no-warnings --experimental-transform-types src/cli.ts",
+    "repl:attach": "node --no-warnings --experimental-transform-types src/cli.ts --attach"
+  },
+  "dependencies": {
+    "@repo/env": "workspace:*",
+    "@repo/interpreter": "workspace:*"
+  },
+  "devDependencies": {
+    "@repo/tsconfigs": "workspace:*",
+    "@types/node": "^22.0.0",
+    "typescript": "^5.9.3",
+    "vitest": "^3.0.0"
+  }
+}
