+++ packages/repl/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@repo/tsconfigs/base.json",
+  "include": ["src", "test", "vitest.config.ts"]
+}
